refactor: firebase encryption
